test should be run in compiler context. move require to AbstractApi so that it can be used in CompilerApi too
