+++ source/class/qx/tool/cli/api/AbstractApi.js
@@ -35,13 +35,53 @@ qx.Class.define("qx.tool.cli.api.AbstractApi", {
   
   members: {
     /**
+     * Loads the configuration data
+     * 
+     * @overridden
+     */
+    async load() {
+      // Nothing
+    },
+    
+      /**
      * Called by the compiler API during initialisation - this is an ideal
      * place to install additional commands, because a command has not yet
      * been selected 
      */
     async initialize() {
       // Nothing
+    },
+
+    /**
+     * 
+     * helper to load an npm module. Check if it can be loaded before
+     * If not install the module with 'npm install --no-save --no-package-lock' to the current library
+     * 
+     * @param module {String} module to check
+     */
+    require: function(module) {
+      let exists = fs.existsSync(path.join(process.cwd(), "node_modules", module));
+      if (!exists) {
+        this.loadNpmModule(module);
+      }      
+      return require(module);
+    },
+    /**
+      * 
+      * install an npm module with 'npm install --no-save --no-package-lock' to the current library
+      * 
+      * @param module {String} module to load
+      */
+    loadNpmModule: function(module) {
+      const {execSync} = require("child_process");
+      let s = `npm install --no-save --no-package-lock ${module}`;
+      qx.tool.compiler.Console.info(s);
+      execSync(s, {
+        stdio: "inherit"
+      });
     }
+	
+
   }
 });
