+++ source/class/qx/tool/cli/api/LibraryApi.js
@@ -39,9 +39,11 @@ qx.Class.define("qx.tool.cli.api.LibraryApi", {
   members: {
     /**
      * Called to load any library-specific configuration and update the compilerConfig
+     * 
+     * @overridden
      */
-    load() {
-      // Nothing
+    async load() {
+      return this.base(arguments);
     },
 
     /**
@@ -51,35 +53,6 @@ qx.Class.define("qx.tool.cli.api.LibraryApi", {
       // Nothing
     },
     
-    /**
-     * 
-     * helper to load an npm module. Check if it can be loaded before
-     * If not install the module with 'npm install --no-save --no-package-lock' to the current library
-     * 
-     * @param module {String} module to check
-     */
-    require: function(module) {
-      let exists = fs.existsSync(path.join(process.cwd(), "node_modules", module));
-      if (!exists) {
-        this.loadNpmModule(module);
-      }      
-      return require(module);
-    },
-    /**
-      * 
-      * install an npm module with 'npm install --no-save --no-package-lock' to the current library
-      * 
-      * @param module {String} module to load
-      */
-    loadNpmModule: function(module) {
-      const {execSync} = require("child_process");
-      let s = `npm install --no-save --no-package-lock ${module}`;
-      qx.tool.compiler.Console.info(s);
-      execSync(s, {
-        stdio: "inherit"
-      });
-    }
-	
     
   }
 });
